Create  Plan for Unsupervised learning model
The fruit data is stored in the fruit_data DataFrame, which has the following structure:

fruit	color	diameter	weight
apple	red	3	150
banana	yellow	2	100
orange	orange	3	125
pear	green	2	95
peach	orange	3	120
kiwi	green	2	90
watermelon	green	4	200
strawberry	red	1	50
raspberry	red	1	55
blueberry	blue	1	60
...	...	...	...
mango	orange	4	150
papaya	orange	4	145
pineapple	yellow	5	250
coconut	brown	5	300
durian	brown	5	275
avocado	green	3	115
fig	brown	2	75
guava	green	2	80
lemon	yellow	3	110
lime	green	2	65
pomegranate	red	4	180
tangerine	orange	2	70
The fruit characteristics (color, diameter, weight) are extracted from the DataFrame and stored in the X variable, while the fruit labels (fruit type) are stored in the y variable. The data is then standardized using the StandardScaler class, and then the KMeans clustering model is trained on the standardized data. The model is then used to predict the cluster assignments for each fruit, and the fruit labels and


The fruit characteristics (color, diameter, weight) are extracted from the DataFrame and stored in the X variable, while the fruit labels (fruit type) are stored in the y variable. The data is then standardized using the StandardScaler class, and then the KMeans clustering model is trained on the standardized data. The model is then used to predict the cluster assignments for each fruit, and the fruit labels and predicted clusters are printed out.
The silhouette score is a measure of how well each sample has been assigned to its own cluster, with a higher score indicating a better assignment. It ranges from -1 to 1, with values closer to 1 indicating a better assignment. The silhouette_score function from the sklearn.metrics module is used to calculate the silhouette score for the given data and predictions. You can use this function or any other appropriate performance metric in the evaluate_model function to evaluate the performance of the model.