Create using reinforcement learning for the  fruit classifier that learns to navigate through a maze by trial and error:
defining a Fruit class that represents a fruit with a name, reward, and location in the maze. It also defines a Maze class that represents the maze and keeps track of the current location of the agent. The QLearning class implements the Q-learning algorithm, and the Trainer class trains the agent by running the Q-learning algorithm for a given number of episodes. The training process involves resetting the maze, selecting an action using the Q-learning algorithm, moving the agent in the maze, and updating the Q-table based on the reward received. Finally, the train method is called to start the training process.