Create Contrastive Language-Image Pre-training model to train a zero-shot classifier.
The script starts by importing the necessary libraries and checking the version of PyTorch. Then, it lists the available models that can be used with CLIP and loads the the dataset to the model.
After loading the model, the script prints out some information about the model, including the number of model parameters, the input resolution (size) of the images that the model expects, the context length (the number of words that the model uses to represent a text), and the size of the vocabulary that the model uses.