Create deploying deep-learning inference networks and deep vision primitives with TensorRT and NVIDIA Jetson.
 the point of the above code is to deploying deep-learning inference networks and deep vision primitives with TensorRT and NVIDIA Jetson.
The Fruit class and main function are placed inside the with statement that creates the CUDA context, so that the context is available to the classify method. The main function creates two instances of the Fruit class and uses the classify method to get the output of the classification for each of them.
after implementing Pre-processing  I will be able to  take and  prepare the input data for the model before it is fed into the TensorRT engine for inference.
In the above code example, the input data is color and size of fruits. The pre-processing step is to format the input data as a numpy array, in this case, the color and size attributes in the Fruit class are passed as a parameter, and they are formatted into a numpy array using np.array([self.color, self.size]).
 Afterward I will be able to implement Post-processing and take steps to interpret and make use of the output from the model after inference.

In the above code example, the output from the model is received as h_output after the context.execute(1, [int(d_input), int(d_output)]) line. There's no post-processing steps defined in the code, it returns h_output as the classification result. Depending on the requirements, you can define post-processing steps as necessary.