Create Vision Transformer in PyTorch for image classification:
n this example, I've added a MultiheadAttention layer and a TransformerEncoder layer to the Vision Transformer. The MultiheadAttention layer takes in the input x, and generates attention weights which are then applied to the input x to generate the output of the attention layer, which is then passed to the transformer encoder. The transformer encoder is composed of multiple layers of a TransformerEncoderLayer. The TransformerEncoder applies a multi-head self-attention mechanism and a position-wise fully connected feed-forward network to the input.
The init_weights function initializes the weights of the model using the Xavier uniform initialization method. The set_trainable function sets all of the model's parameters to be trainable or not trainable. The fine_tune function is used for fine-tuning the model, and allows you to freeze the encoder layers so that they are not trained. The load_state_dict function can be used to load the state dict of the model, which contains the model's parameters and other information.
the Vision Transformer consists of a single transformer encoder and a fully connected layer. The input to the transformer should be of shape (batch_size, sequence_length, d_model), where d_model is the number of features in the input data, and sequence_length is the number of image patches. The output of the transformer is passed through a mean pooling layer to reduce the sequence length and then passed through a fully connected layer to produce the final class scores.